// src/app/interfaces/shipping-app/payment.ts
export interface IPayment {
    shippingBillTo: string;
    shippingBillToDisplayable: string;
    shippingAccountNumber: string;
    shippingAccountNumberDisplayable: string;
    dutiesTaxesBillTo: string;
    dutiesTaxesBillToDisplayable: string;
    dutiesTaxesAccountNumber: string;
    dutiesTaxesAccountNumberDisplayable: string;
}