// src/app/interfaces/shipping-form/billing-form.ts
export interface BillingForm {
    shippingBillTo: string;
    shippingBillToDisplayable: string;
    shippingAccountNumber: string;
    shippingAccountNumberDisplayable: string;
    dutiesTaxesBillTo: string;
    dutiesTaxesBillToDisplayable: string;
    dutiesTaxesAccountNumber: string;
    dutiesTaxesAccountNumberDisplayable: string;
}